New folder w/ project 2 files
The controlling mastermind file is the zach_mastermind.cpp file. It
compiles, but isn't really working yet. I still don't know how to
overload "==" or "=". But the separate header files are working now and
tried to rework the code you gave me for project 2. Too bad you didn't
have that finished to turn in the first time around. Oh well. Still
confused on the increment thing.
To Do:
initialize code class correctly, check the increment function. make the
setCode, getCode, and printCode functions. Check the checkCorrect and
checkIncorrect functions and make sure they work for project 2.
Edit the mastermind file to fit project 2 (started doing this in the
Mastermind_building_from_scratch.cpp file)
complete the setResponse function
edit the main function to work for project 2.

A lot left to do, but did a ton of work... Please look at this and work
on something.