baseline for changing .cpp to .c
We need to change our .cpp files to .c files, but when I did that, I got
an error and the .c file would compile (issues with the #includes).  The
main file would still compile however. We may need to use the .cpp files
for the time being and then switch them before submission.